requirements.txt
Scrapy==1.1.0
awesome-slugify==1.6.5
